Move filter processor to filter_ex due to collision with upstream
also updated devopsext utils and tools are used
