+++ go.mod
@@ -65,8 +65,8 @@ require (
 	github.com/coreos/go-semver v0.3.1
 	github.com/coreos/go-systemd/v22 v22.5.0
 	github.com/couchbase/go-couchbase v0.1.1
-	github.com/devopsext/tools v0.9.13
-	github.com/devopsext/utils v0.2.1
+	github.com/devopsext/tools v0.13.8
+	github.com/devopsext/utils v0.3.3
 	github.com/digitalocean/go-libvirt v0.0.0-20221205150000-2939327a8519
 	github.com/dimchansky/utfbom v1.1.1
 	github.com/djherbis/times v1.6.0
