+++ plugins/processors/all/filter_ex.go
@@ -1,0 +1,5 @@
+//go:build !custom || processors || processors.filter_ex
+
+package all
+
+import _ "github.com/influxdata/telegraf/plugins/processors/filter_ex" // register plugin
